[GPU] Add pattern to fold fill into pad ops
